+++ can/interfaces/__init__.py
@@ -22,6 +22,7 @@ BACKENDS = {
     'neovi':            ('can.interfaces.ics_neovi',        'NeoViBus'),
     'vector':           ('can.interfaces.vector',           'VectorBus'),
     'slcan':            ('can.interfaces.slcan',            'slcanBus'),
+    'canalystii':       ('can.interfaces.canalystii',       'CANalystIIBus'),
     'systec':           ('can.interfaces.systec',           'UcanBus')
 }