+++ can/bus.py
@@ -4,8 +4,6 @@
 Contains the ABC bus implementation and its documentation.
 """
 
-from __future__ import print_function, absolute_import
-
 from abc import ABCMeta, abstractmethod
 import logging
 import threading
@@ -26,7 +24,7 @@ class BusState(Enum):
     ERROR = auto()
 
 
-class BusABC(object):
+class BusABC(metaclass=ABCMeta):
     """The CAN Bus Abstract Base Class that serves as the basis
     for all concrete interfaces.
 
@@ -403,5 +401,3 @@ class BusABC(object):
                  for usage in the interface's bus constructor.
         """
         raise NotImplementedError()
-
-    __metaclass__ = ABCMeta