+++ can/interfaces/kvaser/canlib.py
@@ -8,8 +8,6 @@ specific to Python.
 Copyright (C) 2010 Dynamic Controls
 """
 
-from __future__ import absolute_import
-
 import sys
 import time
 import logging
@@ -69,7 +67,7 @@ class CANLIBError(CanError):
     """
 
     def __init__(self, function, error_code, arguments):
-        super(CANLIBError, self).__init__()
+        super().__init__()
         self.error_code = error_code
         self.function = function
         self.arguments = arguments
@@ -446,7 +444,7 @@ class KvaserBus(BusABC):
         self._timestamp_offset = time.time() - (timer.value * TIMESTAMP_FACTOR)
 
         self._is_filtered = False
-        super(KvaserBus, self).__init__(channel=channel, can_filters=can_filters, **kwargs)
+        super().__init__(channel=channel, can_filters=can_filters, **kwargs)
 
     def _apply_filters(self, filters):
         if filters and len(filters) == 1: