+++ can/interfaces/systec/structures.py
@@ -36,7 +36,7 @@ class CanMsg(Structure):
     ]
 
     def __init__(self, id=0, frame_format=MsgFrameFormat.MSG_FF_STD, data=[]):
-        super(CanMsg, self).__init__(id, frame_format, len(data), (BYTE * 8)(*data), 0)
+        super().__init__(id, frame_format, len(data), (BYTE * 8)(*data), 0)
 
     def __eq__(self, other):
         if not isinstance(other, CanMsg):
@@ -121,8 +121,8 @@ class InitCanParam(Structure):
     ]
 
     def __init__(self, mode, BTR, OCR, AMR, ACR, baudrate, rx_buffer_entries, tx_buffer_entries):
-        super(InitCanParam, self).__init__(sizeof(InitCanParam), mode, BTR >> 8, BTR, OCR, AMR, ACR,
-                                           baudrate, rx_buffer_entries, tx_buffer_entries)
+        super().__init__(sizeof(InitCanParam), mode, BTR >> 8, BTR, OCR, AMR, ACR,
+                        baudrate, rx_buffer_entries, tx_buffer_entries)
 
     def __eq__(self, other):
         if not isinstance(other, InitCanParam):
@@ -273,7 +273,7 @@ class ChannelInfo(Structure):
     ]
 
     def __init__(self):
-        super(ChannelInfo, self).__init__(sizeof(ChannelInfo))
+        super().__init__(sizeof(ChannelInfo))
 
     def __eq__(self, other):
         if not isinstance(other, ChannelInfo):