+++ test/message_helper.py
@@ -5,12 +5,10 @@
 This module contains a helper for writing test cases that need to compare messages.
 """
 
-from __future__ import absolute_import, print_function
-
 from copy import copy
 
 
-class ComparingMessagesTestCase(object):
+class ComparingMessagesTestCase:
     """
     Must be extended by a class also extending a unittest.TestCase.
 
@@ -47,3 +45,13 @@ class ComparingMessagesTestCase(object):
                 print("           message 2: {!r}".format(message_2))
                 self.fail("messages are unequal with allowed timestamp delta {} even when ignoring channels" \
                           .format(self.allowed_timestamp_delta))
+
+    def assertMessagesEqual(self, messages_1, messages_2):
+        """
+        Checks the order and content of the individual messages pairwise.
+        Raises an error if the lengths of the sequences are not equal.
+        """
+        self.assertEqual(len(messages_1), len(messages_2), "the number of messages differs")
+
+        for message_1, message_2 in zip(messages_1, messages_2):
+            self.assertMessageEqual(message_1, message_2)