+++ test/test_load_file_config.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
 # coding: utf-8
+
 import shutil
 import tempfile
 import unittest