+++ doc/history.rst
@@ -46,6 +46,8 @@ The CAN viewer terminal script was contributed by Kristian Sloth Lauszus in 2018
 
 The CANalyst-II interface was contributed by Shaoyu Meng in 2018.
 
+The SYSTEC interface was contributed by @idaniel86
+
 Support for CAN within Python
 -----------------------------